RDKB-33817 : Device was not rebooted when GwProvApp getting Restarted/Stopped
Reason for change: Device was not rebooted when GwProvApp getting Restarted/Stopped
Test Procedure: Restart gwprovapp and check whether box rebooting
                Restart CR and check device getting rebooted
Risks: Low

(cherry picked from commit af71ba682fa20ca0ba71c166e65334474f028d33)

+++ systemd_units/scripts/ProcessResetCheck.sh
@@ -18,22 +18,29 @@
 # limitations under the License.
 #######################################################################################
 # Component Register can safely restart if no other RDKB components have initialized
-
-source /etc/device.properties
-if [ -f /tmp/pam_initialized ] || [ -f /tmp/psm_initialized ]; then
+Set_Reboot_Reason()
+{
 	#This will be launched even during the happy path. For those cases give time for system to shutdown
 	sleep 10
-	if [ -e "/usr/bin/onboarding_log" ]; then
-	    /usr/bin/onboarding_log "Device reboot due to reason CR_crash"
+        if [ -e "/usr/bin/onboarding_log" ]; then
+        	/usr/bin/onboarding_log "Device reboot due to reason $1"
+        fi
+        syscfg set X_RDKCENTRAL-COM_LastRebootReason $1
+        syscfg set X_RDKCENTRAL-COM_LastRebootCounter "1"
+        syscfg commit
+        sync
+        echo "`date`: $2 Crashed Rebooting" >> ${PROCESS_RESTART_LOG}
+        source /rdklogger/logfiles.sh;syncLogs_nvram2
+        reboot
+}
+source /etc/device.properties
+if [ -f "/tmp/CcspCrSsp_Restarted" ];then
+	if [ -f /tmp/pam_initialized ] || [ -f /tmp/psm_initialized ]; then
+		Set_Reboot_Reason "CR_crash" "CcspCrSsp"
+	else
+		echo "`date`: Stopping/Restarting CcspCrSsp" >> ${PROCESS_RESTART_LOG}
+		rm -f /tmp/CcspCrSsp_Restarted
 	fi
-	syscfg set X_RDKCENTRAL-COM_LastRebootReason CR_crash
-	syscfg set X_RDKCENTRAL-COM_LastRebootCounter "1"
-	syscfg commit
-	sync
-	echo "`date`: CcspCrSsp Crashed Rebooting" >> ${PROCESS_RESTART_LOG}
-	source /rdklogger/logfiles.sh;syncLogs_nvram2
-	reboot
-else
-	echo "`date`: Stopping/Restarting CcspCrSsp" >> ${PROCESS_RESTART_LOG}
-	rm -f /tmp/CcspCrSsp_Restarted
+elif [ -f "/tmp/GWPROV_Restarted" ];then
+	Set_Reboot_Reason "GWPROV_crash" "GWPROV"
 fi
